refactor: rename e2e helpers to *.helpers.ts convention
