+++ frontend/e2e/helpers/auth.helpers.ts
