+++ frontend/e2e/helpers/exhaustive.helpers.ts
