+++ frontend/e2e/helpers/table.helpers.ts
