+++ frontend/e2e/test/staff/incidents.spec.ts
@@ -1,21 +1,21 @@
 import { ADMIN_AUTH_FILE } from "../../global-setup";
-import { resetDatabase } from "../../helpers/db";
+import { resetDatabase } from "../../helpers/db.helpers";
 import {
   FilterCase,
   SortCase,
   filterAndExpect,
   filterTestTitle,
   sortAndVerify,
-} from "../../helpers/exhaustive";
-import { expect, suiteTest as test } from "../../helpers/fixtures";
+} from "../../helpers/exhaustive.helpers";
+import { expect, suiteTest as test } from "../../helpers/fixtures.helpers";
 import {
   INCIDENTS,
   LOCATIONS,
   countWhere,
   formatDateInput,
   toTimeString,
-} from "../../helpers/seedData";
-import { Steps } from "../../helpers/steps";
+} from "../../helpers/seed.helpers";
+import { Steps } from "../../helpers/steps.helpers";
 import {
   clickRowAction,
   confirmDialog,
@@ -26,7 +26,7 @@ import {
   selectSidebarCombobox,
   setGlobalSearch,
   waitForTableReady,
-} from "../../helpers/table";
+} from "../../helpers/table.helpers";
 
 // --- Seed data constants -------------------------------------------------------
