+++ frontend/e2e/test/staff/locations.spec.ts
@@ -1,5 +1,5 @@
 import { ADMIN_AUTH_FILE } from "../../global-setup";
-import { resetDatabase } from "../../helpers/db";
+import { resetDatabase } from "../../helpers/db.helpers";
 import {
   FilterCase,
   SortCase,
@@ -7,15 +7,15 @@ import {
   filterTestTitle,
   openIncidentSidebar,
   sortAndVerify,
-} from "../../helpers/exhaustive";
-import { expect, suiteTest as test } from "../../helpers/fixtures";
+} from "../../helpers/exhaustive.helpers";
+import { expect, suiteTest as test } from "../../helpers/fixtures.helpers";
 import {
   LOCATIONS,
   countWhere,
   firstUniqueToken,
   formatDateInput,
-} from "../../helpers/seedData";
-import { Steps } from "../../helpers/steps";
+} from "../../helpers/seed.helpers";
+import { Steps } from "../../helpers/steps.helpers";
 import {
   clickRowAction,
   confirmDialog,
@@ -29,7 +29,7 @@ import {
   setGlobalSearch,
   setPageSize,
   waitForTableReady,
-} from "../../helpers/table";
+} from "../../helpers/table.helpers";
 
 // --- Seed data constants -------------------------------------------------------
