+++ frontend/e2e/test/staff/students.spec.ts
@@ -1,13 +1,13 @@
 import { ADMIN_AUTH_FILE } from "../../global-setup";
-import { resetDatabase } from "../../helpers/db";
+import { resetDatabase } from "../../helpers/db.helpers";
 import {
   FilterCase,
   SortCase,
   filterAndExpect,
   filterTestTitle,
   sortAndVerify,
-} from "../../helpers/exhaustive";
-import { expect, suiteTest as test } from "../../helpers/fixtures";
+} from "../../helpers/exhaustive.helpers";
+import { expect, suiteTest as test } from "../../helpers/fixtures.helpers";
 import {
   PARTIES,
   STUDENTS,
@@ -16,8 +16,8 @@ import {
   countWhere,
   currentAcademicYearStart,
   firstUniqueToken,
-} from "../../helpers/seedData";
-import { Steps } from "../../helpers/steps";
+} from "../../helpers/seed.helpers";
+import { Steps } from "../../helpers/steps.helpers";
 import {
   clickRowAction,
   confirmDialog,
@@ -27,7 +27,7 @@ import {
   selectSidebarCombobox,
   setGlobalSearch,
   waitForTableReady,
-} from "../../helpers/table";
+} from "../../helpers/table.helpers";
 
 // --- Seed data constants -------------------------------------------------------
